+++ src/pool/_prep.js
@@ -24,7 +24,7 @@ const validSkipNsfwOption = function (option) {
 
 // throw early on bad options. the pool turns this into an 'error' event and a rejected `done`
 const checkOptions = function (opts) {
-  const { file, format, namespace, skip_redirect, skip_disambig, skip_nsfw, skip_stub } = opts
+  const { file, format, namespace, skip_redirect, skip_disambig, skip_nsfw, skip_stub, silent } = opts
   const renamed = Object.keys(renamedFilters).find((name) => Object.hasOwn(opts, name))
   if (renamed) {
     throw new Error(`'${renamed}' has been renamed to '${renamedFilters[renamed]}' with inverted boolean semantics`)
@@ -52,6 +52,9 @@ const checkOptions = function (opts) {
   }
   if (typeof skip_stub !== 'boolean') {
     throw new Error(`'skip_stub' must be true or false`)
+  }
+  if (typeof silent !== 'boolean') {
+    throw new Error(`'silent' must be true or false`)
   }
 }
 export { checkOptions, formats }